fix: remove cyclic dependencies
