+++ packages/actions-kit/package.json
@@ -91,12 +91,12 @@
   },
   "dependencies": {
     "zod": "catalog:",
-    "@actions-kit/action-schema": "workspace:*",
+    "@actions-kit/config": "workspace:*",
     "cac": "catalog:",
     "consola": "catalog:",
-    "c12": "catalog:",
     "defu": "catalog:",
     "unplugin-actions-kit": "workspace:*",
+    "@actions-kit/action-schema": "workspace:*",
     "js-yaml": "catalog:"
   },
   "devDependencies": {
