+++ packages/builders/esbuild/esbuild.d.ts
@@ -1,4 +1,4 @@
-declare module "actions-kit/config" {
+declare module "@actions-kit/config" {
 	import type { BuildOptions as ESBuildBuildOptions } from "esbuild";
 
 	interface ActionsKitConfig {
