+++ packages/builders/esbuild/src/index.ts
@@ -1,4 +1,4 @@
-import type { ActionsKitConfig } from "actions-kit/config";
+import type { ActionsKitConfig } from "@actions-kit/config";
 import { defu } from "defu";
 import type { BuildOptions as ESBuildBuildOptions } from "esbuild";
 import { getESBuildEntryPoint } from "./utils";
