+++ packages/builders/rolldown/rolldown.d.ts
@@ -1,4 +1,4 @@
-declare module "actions-kit/config" {
+declare module "@actions-kit/config" {
 	import type { RolldownOptions } from "rolldown";
 
 	interface ActionsKitConfig {
