+++ packages/builders/rolldown/src/index.ts
@@ -1,4 +1,4 @@
-import type { ActionsKitConfig } from "actions-kit/config";
+import type { ActionsKitConfig } from "@actions-kit/config";
 import { defu } from "defu";
 import type { RolldownOptions } from "rolldown";
 import { join } from "node:path";
