+++ packages/builders/rollup/rollup.d.ts
@@ -1,4 +1,4 @@
-declare module "actions-kit/config" {
+declare module "@actions-kit/config" {
 	import type { RollupOptions } from "rollup";
 
 	interface ActionsKitConfig {
