+++ packages/builders/vite/src/index.ts
@@ -1,4 +1,4 @@
-import type { ActionsKitConfig } from "actions-kit/config";
+import type { ActionsKitConfig } from "@actions-kit/config";
 import { defu } from "defu";
 import type { InlineConfig } from "vite";
 import { join, resolve } from "node:path";
