+++ packages/builders/vite/vite.d.ts
@@ -1,4 +1,4 @@
-declare module "actions-kit/config" {
+declare module "@actions-kit/config" {
 	import type { InlineConfig } from "vite";
 
 	interface ActionsKitConfig {
