+++ packages/config/README.md
@@ -1,0 +1,3 @@
+# @actions-kit/config
+
+Configuration utilities for Actions Kit.
