+++ packages/config/tsconfig.build.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["../../tsconfig.base.build.json"],
+  "include": ["src/**/*.ts"]
+}
